Move Mapping so getObjectMapper can be hidden
