+++ src/main/java/org/jongo/marshall/jackson/Mapping.java
@@ -14,7 +14,10 @@
  * limitations under the License.
  */
 
-package org.jongo.marshall.jackson.configuration;
+package org.jongo.marshall.jackson;
+
+import org.jongo.marshall.jackson.configuration.ReaderCallback;
+import org.jongo.marshall.jackson.configuration.WriterCallback;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectReader;
@@ -32,7 +35,7 @@ public class Mapping {
         this.writerCallback = writerCallback;
     }
 
-    public ObjectMapper getObjectMapper() {
+    ObjectMapper getObjectMapper() {
         return mapper;
     }
