+++ src/test/java/org/jongo/marshall/jackson/ConfigurationHelper.java
@@ -1,6 +1,5 @@
 package org.jongo.marshall.jackson;
 
-import org.jongo.marshall.jackson.configuration.Mapping;
 
 public class ConfigurationHelper {
